♻️ Refactor: #392 Extract shared texture loading hook
