+++ apps/scene-studio/src/primitives/DepthGallery/ImagePlane.tsx
@@ -1,12 +1,4 @@
-import { useThree } from '@react-three/fiber';
-import { useEffect, useState } from 'react';
-import {
-  cancelRender,
-  continueRender,
-  delayRender,
-  useRemotionEnvironment,
-} from 'remotion';
-import { SRGBColorSpace, type Texture, TextureLoader } from 'three';
+import { useImageTexture } from '../shared/useImageTexture';
 
 interface Props {
   src: string;
@@ -19,37 +11,7 @@ const PLANE_WIDTH_IN_WORLD_UNITS = 2;
 const PLANE_HEIGHT_IN_WORLD_UNITS = 3;
 
 export function ImagePlane({ src, position, opacity }: Props) {
-  const [texture, setTexture] = useState<Texture | null>(null);
-  const [delayRenderHandle] = useState(() =>
-    delayRender(`Loading DepthGallery texture: ${src.slice(0, 48)}`)
-  );
-  const advance = useThree((state) => state.advance);
-  const { isRendering } = useRemotionEnvironment();
-
-  useEffect(() => {
-    new TextureLoader().load(
-      src,
-      (loadedTexture) => {
-        loadedTexture.colorSpace = SRGBColorSpace;
-        setTexture(loadedTexture);
-      },
-      undefined,
-      (error) => cancelRender(error)
-    );
-  }, [src]);
-
-  useEffect(() => {
-    if (!texture) {
-      return;
-    }
-    // During rendering the frameloop is 'never' and the canvas was already
-    // drawn for this frame before the texture arrived; advance once so the
-    // screenshot includes the textured mesh, then release Remotion.
-    if (isRendering) {
-      advance(performance.now());
-    }
-    continueRender(delayRenderHandle);
-  }, [texture, isRendering, advance, delayRenderHandle]);
+  const texture = useImageTexture(src);
 
   if (!texture) {
     return null;
